em casa

+++ libft/ft_memcmp.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memcmp.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kyoshi <kyoshi@student.42.fr>              +#+  +:+       +#+        */
+/*   By: kakubo-l <kakubo-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/16 20:23:29 by kyoshi            #+#    #+#             */
-/*   Updated: 2025/07/16 20:26:01 by kyoshi           ###   ########.fr       */
+/*   Updated: 2025/07/22 19:54:24 by kakubo-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	unsigned char		*ptr;
 	unsigned char		*ptr1;
-	int					i;
+	size_t				i;
 
 	i = 0;
 	if (n == 0)
